+++ common/src/main/java/io/github/flemmli97/runecraftory/common/items/creative/ItemDebug.java
@@ -1,8 +1,5 @@
 package io.github.flemmli97.runecraftory.common.items.creative;
 
-import io.github.flemmli97.runecraftory.common.world.GateSpawning;
-import net.minecraft.Util;
-import net.minecraft.network.chat.TextComponent;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResultHolder;
@@ -10,9 +7,6 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.level.levelgen.feature.ConfiguredStructureFeature;
-
-import java.util.Set;
 
 public class ItemDebug extends Item {
 
@@ -23,7 +17,7 @@ public class ItemDebug extends Item {
     @Override
     public InteractionResultHolder<ItemStack> use(Level level, Player player, InteractionHand handIn) {
         if (level instanceof ServerLevel serverLevel) {
-            long time = System.nanoTime();
+            /*long time = System.nanoTime();
             Set<ConfiguredStructureFeature<?, ?>> structures = GateSpawning.getStructuresAt(serverLevel, player.blockPosition());
             long delta = System.nanoTime() - time;
             player.sendMessage(new TextComponent("" + structures), Util.NIL_UUID);